Add farmer edit password layout
